Ticket 11 Delete Comments
Created Authentication for user and updated ticket 8 so now users can only comment if logged in and implemented functioanlity to only allow comments to be deleted if the user logged in matches the author. Moved commented and deleted states so the article comment count also  updates on a post or delete.